fix(awards): make the criteria builder usable
Found by walking the builder in a browser; each of these made a rule
silently wrong or the page unusable.

- award_rules.award_id was unsignedBigInteger against awards.id
  increments(), which MySQL 8 rejects. The migration threw and took all
  125 awards tests with it; SQLite and PostgreSQL both hid it
- livewire payload nesting raised to 40. One OR group around one rule is
  11 levels, so OR groups could not be submitted at all
- the aggregate and comparison selects render a default with no
  placeholder, so required() rejected a field that visibly had a value
- a stored tree the builder cannot read now opens empty with a warning
  rather than fataling the edit page, which left the award unrepairable.
  Import refuses such a tree outright instead of storing it
- airport columns are searchable pickers keyed on the model key, so they
  survive airport ids becoming ULIDs, and a missing airport no longer
  drops the criterion (which would grant the award to everyone)
- form actions sit in the last section footer, primary furthest right,
  with tabler icons; run test joins them and reports who matches

+++ database/migrations/2026_08_08_062734_create_award_rules_and_snippets.php
@@ -34,7 +34,9 @@ return new class() extends Migration
                 $table->charset = 'utf8mb4';
 
                 $table->id();
-                $table->unsignedBigInteger('award_id')->unique();
+                // `awards.id` is `increments()` (int unsigned), not `id()` --
+                // the types must match exactly or MySQL 8 rejects the FK below.
+                $table->unsignedInteger('award_id')->unique();
                 $table->json('conditions');
                 $table->timestamps();
